move impl_expected_batch_write_hook macro
