+++ open-coroutine-hooks/Cargo.toml
@@ -11,9 +11,9 @@ license = "LGPL-3.0 OR Apache-2.0"
 readme = "../README.md"
 
 [dependencies]
-libc = "0.2.138"
-once_cell = "1.13.0"
-num_cpus = "1.14.0"
+libc = "0.2.150"
+once_cell = "1.18.0"
+num_cpus = "1.16.0"
 cfg-if = "1.0.0"
 open-coroutine-core = { version = "0.5.0", path = "../open-coroutine-core" }
 open-coroutine-timer = { version = "0.5.0", path = "../open-coroutine-timer" }
